requirements.txt
PyYAML==6.0
numpy==1.23.4
pandas==1.5.1
scikit-learn==1.1.3
click==8.1.3